% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annot.R
\name{add_annot}
\alias{add_annot}
\title{Annotate spectra}
\usage{
add_annot(annobject, ppm = 5)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}

\item{ppm}{mass deviation for elemental composition
search in ppm}
}
\value{
Annotate spectra using the compound elemental composition
with \code{Spec2Annot} and return an annotated annobject.
}
\description{
Annotate spectra
}
