% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{add_best_peaks}
\alias{add_best_peaks}
\title{Add peaks detection to annobject}
\usage{
add_best_peaks(annobject, minscan = 3, rt_limit = NULL, debug = FALSE)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}

\item{minscan}{numeric value for the minimum number of scan
inside a peak to keep it}

\item{rt_limit}{limit rt deviation from reference when searching peaks}

\item{debug}{Logical to show log messages (\code{TRUE}) or not (\code{FALSE})}
}
\value{
The annobject with the Peaks slots for each compound
filled.
}
\description{
Add peaks detection to annobject
}
