% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpd_add_events.R
\name{add_cpd_events}
\alias{add_cpd_events}
\title{Add MSn events for each compounds}
\usage{
add_cpd_events(annobject, prec_ppm = 10)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}

\item{prec_ppm}{Mass tolerance to find the precursor (in ppm)}
}
\value{
Fill the \code{MSEvents} compound slot of the annobject
}
\description{
Add MSn events for each compounds
}
