% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_events.R
\name{add_events}
\alias{add_events}
\title{Add MSn events}
\usage{
add_events(annobject, firstevent = TRUE)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}

\item{firstevent}{Should all MSn events be checked individually
(\code{FALSE}) or only the first one (\code{TRUE}).}
}
\value{
Implement the \code{MSEvents} and \code{info} slot of the object
}
\description{
Add MSn events
}
