% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_events_to_extract.R
\name{add_events_to_extract}
\alias{add_events_to_extract}
\title{Add MS(n) events to extract}
\usage{
add_events_to_extract(annobject, debug = FALSE)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}

\item{debug}{Logical to show log messages (\code{TRUE}) or not (\code{FALSE})}
}
\value{
Fill the \code{MSspectra} slots of the annobject with
information on MS(n) event to extract in the next
step.
}
\description{
Add MS(n) events to extract
}
