% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mspurity.R
\name{add_mspurity}
\alias{add_mspurity}
\title{Add MS(n) purity}
\usage{
add_mspurity(annobject, debug = FALSE)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}

\item{debug}{Logical to show log messages (\code{TRUE}) or not (\code{FALSE})}
}
\value{
Fill the \code{IsoPurity} slot of annobject
}
\description{
Add MS(n) purity
}
