% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spectra.R
\name{add_spectra}
\alias{add_spectra}
\title{Add spectrum to annobject}
\usage{
add_spectra(annobject)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}
}
\value{
Fill the \code{spectra} slots of annobject.
}
\description{
Add spectrum to annobject
}
