% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{calculate_zigzag}
\alias{calculate_zigzag}
\title{Calculate the zig-zag score of a vector}
\source{
\url{https://github.com/GauravPandeyLab/MetaClean/blob/master/R/calculateZigZagIndex.R}
}
\usage{
calculate_zigzag(vecint)
}
\arguments{
\item{vecint}{Numeric vector containing the intensities
sorted by the time dimension.}
}
\value{
The zig-zag index value
}
\description{
Calculate the zig-zag score of a vector
}
\examples{
calculate_zigzag(rnorm(100, 1e5, sd = 1e3))
calculate_zigzag(rnorm(100, 1e5, sd = 1e3))
}
