% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpd_add_events.R
\name{cpd_add_events}
\alias{cpd_add_events}
\title{Get MS2 events for each compounds}
\usage{
cpd_add_events(cpd_dt, ms_events_dt, prec_ppm = 10, indexonly = FALSE)
}
\arguments{
\item{cpd_dt}{Compound table as returned by \code{Spec2Xtract::parse_index_dt()}}

\item{ms_events_dt}{data.table with MS events found in
\code{annobject$file$MSEvents}}

\item{prec_ppm}{Mass tolerance to find the precursor (in ppm)}

\item{indexonly}{Logical to return MS events info (\code{FALSE}) or ony
the index (\code{TRUE})}
}
\value{
A data.table containing the MS(n) events related to the compound.
If \code{indexonly = TRUE} only the MS(n) events indexes are returned,
else all the available informations are returned.
}
\description{
Get MS2 events for each compounds
}
