% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{cpd_add_ionsmass}
\alias{cpd_add_ionsmass}
\title{Calculate ion mass from elemental composition}
\usage{
cpd_add_ionsmass(cpd)
}
\arguments{
\item{cpd}{A table containing compound information, must have the following
columns: compound, rtsec, elemcomposition and optionally inchikey}
}
\value{
Return the \code{cpd} table with:
mz_neutral, mz_pos and mz_neg values
calculated from the elemental composition
}
\description{
Calculate ion mass from elemental composition
}
