% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{detect_centwave_peaks}
\alias{detect_centwave_peaks}
\title{Use MassSpecWavelet to detect peaks}
\usage{
detect_centwave_peaks(
  vecint,
  snrth = 3,
  majpeakonly = TRUE,
  limits = c("xic", "scales")[2],
  ...
)
}
\arguments{
\item{vecint}{Numeric vector containing the intensities
sorted by the time dimension.}

\item{snrth}{Signal over Noise threshold (see
\code{MassSpecWavelet::peakDetectionCWT()})}

\item{majpeakonly}{Logical to extract only the majors peaks (\code{TRUE})
or all peaks (\code{FALSE}) from MassSpecWavelet}

\item{limits}{Should peak boundaries be found on true data (\code{xic}) or
on CentWave scales (\code{scales}: default)?}

\item{...}{Other options passed to \code{MassSpecWavelet::peakDetectionCWT()}}
}
\value{
A data.table with information on chromatographic peaks
with the following columns:
\itemize{
\item \code{peakID}: Peak name given by \code{MassSpecWavelet::peakDetectionCWT()}
\item \code{scmin}, \code{scmax}, \code{scpos}: left, right boundaries and apex of the peak
(index of the values in vecint)
\item \code{scale}: scale of the peak given by \code{MassSpecWavelet::peakDetectionCWT()}
}
}
\description{
Use MassSpecWavelet to detect peaks
}
