% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table.R
\name{export_spectrum_table}
\alias{export_spectrum_table}
\title{Write spectrum to table}
\usage{
export_spectrum_table(annobject, save_dir)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}

\item{save_dir}{Directory to save each spectrum}
}
\description{
Write spectrum to table
}
