% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table.R
\name{export_tables}
\alias{export_tables}
\title{Export spectra as tables}
\usage{
export_tables(annobject, save_dir)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}

\item{save_dir}{Directory to save tables}
}
\description{
Export spectra as tables
}
