% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{fun_best_peaks}
\alias{fun_best_peaks}
\title{Return best peak}
\usage{
fun_best_peaks(pk_scores)
}
\arguments{
\item{pk_scores}{A data.table containing peaks info and scores}
}
\value{
A data.table containing the only best peaks for each \code{CpdIndex}
based on the maximum of \code{peak_score} column
}
\description{
Return best peak
}
