% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{fun_check_cpd}
\alias{fun_check_cpd}
\title{Check compound file}
\usage{
fun_check_cpd(cpd)
}
\arguments{
\item{cpd}{A table containing compound information, must have the
following columns: \code{compound}, \code{rtsec}, \code{elemcomposition} and
optionally \code{inchikey}}
}
\value{
Return a formated \code{data.table} with input informations +
the following columns:
\itemize{
\item \code{CpdIndex}: unique \code{integer} to identify the compound entry.
\item \code{rtmin}: retention time in minutes.
}
}
\description{
Check compound file
}
\examples{
fun_check_cpd(Spec2Xtract:::example_cpdlist_realdt)

}
