% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_xics.R
\name{fun_get_xic}
\alias{fun_get_xic}
\title{Extract a single XIC}
\usage{
fun_get_xic(rawpath, mz, ppm, filter, type = c("xic", "tic", "bpc")[1])
}
\arguments{
\item{rawpath}{Path to the \code{.raw} file}

\item{filter}{special string to filter events
(ex.: \code{"ms"}, or \code{"ms2"})}
}
\value{
A \code{data.table} containing the XIC with the following columns:
\code{filter}: the scanType field
\code{ppm}: the ppm parameter used
\code{mass}: the center mass of the XIC
\code{rt}: the retention time
\code{i}: the intensities
}
\description{
Extract a single XIC
}
\examples{
fun_get_xic(
  rawpath = rawrr::sampleFilePath(),
  filter = "ms2",
  mz = 669.8381,
  ppm = 10,
  type = "xic"
)

}
