% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_xics.R
\name{fun_get_xics}
\alias{fun_get_xics}
\title{Extract multiple XICs}
\usage{
fun_get_xics(
  rawpath,
  mz = NULL,
  ppm = 10,
  mslevel = 1L,
  isomz = NULL,
  index_table = NULL,
  type = c("xic", "tic")[1]
)
}
\arguments{
\item{rawpath}{Path to the .raw file}

\item{mz}{centered mass of the XIC window}

\item{ppm}{mass range of the XIC window (in ppm)}

\item{mslevel}{MS(level) to extract as integer}

\item{isomz}{(optional) mass of the isolation window to extract}

\item{index_table}{The index table as returned by
\code{Spec2Xtract::get_rawindex()}}

\item{type}{string to extract either the \code{"xic"} or \code{"tic"}}
}
\value{
a data.table with the XICs
}
\description{
Extract multiple XICs
}
