% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table.R
\name{get_event_summary}
\alias{get_event_summary}
\title{Title}
\usage{
get_event_summary(annobject)
}
\arguments{
\item{annobject}{Object build with \code{Spec2Extract::init_object()}}
}
\description{
Title
}
