% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_events_to_extract.R
\name{get_events_scans}
\alias{get_events_scans}
\title{Get events}
\usage{
get_events_scans(cpd_events, pk_best, index_table)
}
\arguments{
\item{cpd_events}{Compound MS(n) event slot}

\item{pk_best}{data.table with details on the best peak}

\item{index_table}{The index table as returned by
\code{Spec2Xtract::get_rawindex()}}
}
\value{
a data.table with details on the MS(n) events to extract
for each compound.
}
\description{
Get events
}
