% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{get_peak_range}
\alias{get_peak_range}
\title{Use MassSpecWavelet to detect peaks}
\usage{
get_peak_range(xic, direction = c("R", "L"), peak_sel_center)
}
\arguments{
\item{xic}{The XIC as a data.table with the following columns:
\itemize{
\item \code{scan}: scan index as unique \code{integer}
\item \code{i}: intensities as \code{numeric}
}}

\item{direction}{Character to choose which side of the peak to return:
the right 'R' or left 'L' side}

\item{peak_sel_center}{Scan corresponding to the peak apex (starting point).}
}
\value{
A numeric value corresponding to the scan on the right \code{R} or
left \code{L} side of the peak.
}
\description{
Use MassSpecWavelet to detect peaks
}
