% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{get_peak_scores_mult}
\alias{get_peak_scores_mult}
\title{Calculate peak scores}
\usage{
get_peak_scores_mult(xic_peaks, cpd_events, CpdIndex = NULL, cpd_info = NULL)
}
\arguments{
\item{xic_peaks}{Peak list as returned by
\code{Spec2Xtract::get_peaks_from_xic()}}

\item{cpd_events}{Compound MS(n) event slot}

\item{CpdIndex}{(optional) Compound iteration ID to use}

\item{cpd_info}{Compound Info slot or reference rt in minute}
}
\value{
The \code{xic_peaks} data.table with a new columns:
\itemize{
\item \code{peak_score}: Overall scoring of the peaks
\item \code{peak_in_range}: Number of peaks in range
\item \code{peaks_mslevels}: MS(n) level of the peak
\item \code{diff_from_ref}: Time difference from reference
}
}
\description{
Calculate peak scores
}
