% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{get_peaks_from_cpdsXIC}
\alias{get_peaks_from_cpdsXIC}
\title{Extract peaks from XICs}
\usage{
get_peaks_from_cpdsXIC(xic_cpd, minscan = 4)
}
\arguments{
\item{xic_cpd}{a data.table with XIC informations:
columns: \code{filter}, \code{MSEvent_index}, \code{rt}
\code{i}}

\item{minscan}{numeric value for the minimum number of scan
inside a peak to keep it}
}
\description{
Extract peaks from XICs
}
