% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{get_peaks_from_xic}
\alias{get_peaks_from_xic}
\title{Extract peaks from XICs}
\usage{
get_peaks_from_xic(xic_cpd, minscan = 4)
}
\arguments{
\item{xic_cpd}{a data.table with XIC informations:
columns: \code{filter}, \code{EventIndex}, \code{rt}
\code{i}}

\item{minscan}{Threshold for the minimum number of scan inside a peak to keep it as \code{integer}.}
}
\description{
Extract peaks from XICs
}
