% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{get_peaks_xic}
\alias{get_peaks_xic}
\title{Find peaks usint CentWave}
\usage{
get_peaks_xic(xic_mat, minscan = 3, majpeakonly = FALSE)
}
\arguments{
\item{xic_mat}{A \code{matrix} or \code{data.table} with the following columns:
\itemize{
\item \code{rt} with the retention time as \code{numeric}
\item \code{i} with the intensities as \code{numeric}
}}

\item{minscan}{Threshold for the minimum number of scan inside a peak to keep it as \code{integer}.}

\item{majpeakonly}{Logical to extract only the majors peaks (\code{TRUE})
or all peaks (\code{FALSE}) with
\code{\link[MassSpecWavelet]{peakDetectionCWT}}}
}
\description{
Find peaks usint CentWave
}
