% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{get_rawindex}
\alias{get_rawindex}
\title{Get .raw index as data.table}
\usage{
get_rawindex(rawpath)
}
\arguments{
\item{rawpath}{Path to the \code{.raw} file}
}
\value{
Return a data.table with the spectra indexes from \code{rawrr::readIndex}.
}
\description{
Get .raw index as data.table
}
\examples{
get_rawindex(rawrr::sampleFilePath())
}
