% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spectra.R
\name{get_spectrum_db}
\alias{get_spectrum_db}
\title{Extract spectrum}
\usage{
get_spectrum_db(ms_scan_to_get, rawpath)
}
\arguments{
\item{ms_scan_to_get}{scan index of the spectrum to extract}

\item{rawpath}{Path to the .raw file}
}
\value{
Return a list with two levels:
\itemize{
\item \code{spectra_info_dt}: a data.table containing details on each
spectrum.
\item \code{spectra_db}: a data.table containing the spectrum (mz, i).
}
}
\description{
Extract spectrum
}
\examples{
get_spectrum_db(12, rawrr::sampleFilePath())

}
