% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{init_object}
\alias{init_object}
\title{Initialize object to store results}
\usage{
init_object(files, cpd)
}
\arguments{
\item{files}{Vector containing the paths to the .raw or .mzML files}

\item{cpd}{A table containing compound information, must have the following
columns: compound, rtsec, elemcomposition and optionally inchikey}
}
\value{
Return a list with the following levels:
\itemize{
\item \code{"cpd_info"}: data.table with information on the compounds from cpd.
\item \code{"XICs"}: level to store XICs informations
\item \code{"Peaks"}: level to store peaks informations
\item \code{"MSspectra"}: level to store mass spectrum
}
}
\description{
Initialize object to store results
}
