% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{load_files}
\alias{load_files}
\title{Create data.table from files path}
\usage{
load_files(files)
}
\arguments{
\item{files}{Vector of path(s) to \code{.raw} files}
}
\value{
Details on returned object.
}
\description{
Description
}
\examples{
load_files(files = rawrr::sampleFilePath())
}
