% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{open_cpd_file}
\alias{open_cpd_file}
\title{Open a table in different format}
\usage{
open_cpd_file(cpd_path)
}
\arguments{
\item{cpd_path}{path to the table with compound informations}
}
\value{
The compound table in \code{data.table} format.
}
\description{
Open a table in one of \code{.xlsx}, \code{.csv}, \code{.tsv}, \code{.txt}
formats and convert to a data.table
}
