% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spectra.R
\name{parse_spectrum_object}
\alias{parse_spectrum_object}
\title{Parse a spectrum object}
\usage{
parse_spectrum_object(spectrum_list, mz_range = NULL, mslevel = NULL)
}
\arguments{
\item{spectrum_list}{list of spectra}

\item{mz_range}{(optional) mass range to keep}

\item{mslevel}{Not sure DEV}
}
\value{
A list of spectra as \code{spectrum_list} but with
formatted column and types.
}
\description{
Parse a spectrum object
}
