% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_Spec2Xtract.R
\name{run_Spec2Xtract}
\alias{run_Spec2Xtract}
\title{Wrapper to run targets pipeline}
\usage{
run_Spec2Xtract(
  files_dir,
  cpd_path,
  firstevent = TRUE,
  prec_ppm = 5,
  minscan = 3,
  rt_limit = 0.2,
  ppm = 5,
  filter_irel = 0,
  filter_isopurity = 0,
  one_msp_file = TRUE,
  save_dir = "./report",
  ncore = 1
)
}
\arguments{
\item{files_dir}{path to the directory containing the
.raw files}

\item{cpd_path}{path to the table with compound informations}

\item{filter_irel}{Filter ions based on relative intensity}

\item{filter_isopurity}{Filter ions based on the isolation purity (from 0 to 100)}

\item{one_msp_file}{Boolean to combine all spectra to one unique .msp file \code{TRUE} or to
separate files \code{FALSE}}

\item{ncore}{Number of parallel workers}
}
\description{
This function run a full targets pipeline from
two paths: one to the directory containing .raw
files and one to the compound table.
}
