% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_msp.R
\name{save_as_msp}
\alias{save_as_msp}
\title{Save spectrum as .msp file}
\usage{
save_as_msp(spectra_dt, spectra_info, cpd_info, file_out)
}
\arguments{
\item{spectra_dt}{Spectrum as data.table with columns
(mz, irel, formula)}

\item{spectra_info}{Spectrum info to add}

\item{cpd_info}{Compound info to add}

\item{file_out}{Output path (example: \verb{/path/to/file.msp})
if \code{NULL} the function will return a list with the file
content.}
}
\value{
Save the .msp file to the file_out directory and
return a list with the .msp content
}
\description{
Save spectrum as .msp file
}
