% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spec2Xtract-package.R
\docType{package}
\name{Spec2Xtract-package}
\alias{Spec2Xtract}
\alias{Spec2Xtract-package}
\title{Spec2Xtract: Extract MS2 spectra from LC-MS}
\description{
Extract MS2 spectra from .raw Thermo files acquired in LC-MS/MS using the rawrr package, control chromatogram and spectral quality and return the MS2 spectra in different format (table, msp).
}
\author{
\strong{Maintainer}: Sylvain Dechaumet \email{sylvain.dechaumet@cea.fr}

Other contributors:
\itemize{
  \item Etienne Thevenot \email{etienne.thevenot@cea.fr} [contributor]
  \item Eric Venot \email{eric.venot@cea.fr} [reviewer]
  \item Annelaure Damont \email{annelaure.damont@cea.fr} [contributor]
  \item Chenqin Cao \email{chenqin.cao@cea.fr} [data contributor]
  \item Anaïs Legrand \email{anais.legrand@cea.fr} [contributor]
}

}
\keyword{internal}
