% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_msp.R
\name{write_file}
\alias{write_file}
\title{Write file to disk}
\usage{
write_file(string, file_path, overwrite = TRUE)
}
\arguments{
\item{string}{string as vector or list to write to the file}

\item{file_path}{Path to the output file}

\item{overwrite}{Boolean to append \code{FALSE} or overwrite
the file \code{TRUE}}
}
\value{
A Boolean if the file was written \code{TRUE} or not \code{FALSE}
}
\description{
Write file to disk
}
