% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{xtract_spectra}
\alias{xtract_spectra}
\title{Wrapper to detect and extract spectra}
\usage{
xtract_spectra(
  files,
  cpd,
  firstevent = TRUE,
  prec_ppm = 10,
  ms1 = TRUE,
  rt_limit = 0.15,
  ppm = 5,
  save_dir = NULL,
  debug = FALSE
)
}
\arguments{
\item{files}{Vector containing the paths to the .raw or .mzML files}

\item{cpd}{A table containing compound information, must have the following
columns: compound, rtsec, elemcomposition and optionally inchikey}

\item{firstevent}{Should all MSn events be checked individually
(\code{FALSE}) or only the first one (\code{TRUE}).}

\item{prec_ppm}{Mass tolerance to find the precursor (in ppm)}

\item{ms1}{Logical}

\item{rt_limit}{limit rt deviation from reference when searching peaks (in seconds)}

\item{ppm}{mass deviation for elemental composition
search in ppm}

\item{save_dir}{Directory to save tables}

\item{debug}{Logical to show log messages (\code{TRUE}) or not (\code{FALSE})}
}
\value{
If save_dir is not null, export the spectra and summary there.
Return a list with the following levels:
\itemize{
\item \code{"cpd_info"}: data.table with information on the compounds from cpd.
\item \code{"XICs"}: level to store XICs informations
\item \code{"Peaks"}: level to store peaks informations
\item \code{"MSspectra"}: level to store mass spectrum
}
}
\description{
Wrapper to detect and extract spectra
}
