% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfile.R
\name{check_rawfile}
\alias{check_rawfile}
\title{Check if raw file exist}
\usage{
check_rawfile(rawpath)
}
\arguments{
\item{rawpath}{Path to the .raw file}
}
\value{
This function check .raw and .mzML files for readability.
If the file is readable it will return the full path, if
not it will stop with an error.
}
\description{
Check if raw file exist
}
\examples{
check_rawfile(rawrr::sampleFilePath())
}
