% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{cpd_add_ionsmass}
\alias{cpd_add_ionsmass}
\title{Calculate ion mass from the elemental composition}
\usage{
cpd_add_ionsmass(cpd)
}
\arguments{
\item{cpd}{A \code{data.table} produced by from \code{\link[=fun_check_cpd]{fun_check_cpd()}}}
}
\value{
Return a \code{data.table} with the following columns:
\itemize{
\item \code{mz_neutral}: neutral mass
\item \code{mz_pos}: positive mass
\item \code{mz_neg}: negative mass
}
}
\description{
Use \link{Spec2Annot} functions to calculate the ions masses
from the neutral elemental composition.
}
