% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{detect_centwave_peaks}
\alias{detect_centwave_peaks}
\title{Use MassSpecWavelet to detect peaks}
\usage{
detect_centwave_peaks(
  vecint,
  SNR.Th = 3,
  majpeakonly = TRUE,
  limits = c("xic", "scales")[2],
  ...
)
}
\arguments{
\item{vecint}{Intensities
sorted by the time dimension as a \verb{numeric vector}}

\item{SNR.Th}{SNR (Signal to Noise Ratio) threshold}

\item{majpeakonly}{Logical to extract only the majors peaks (\code{TRUE})
or all peaks (\code{FALSE}) with
\code{\link[MassSpecWavelet]{peakDetectionCWT}}}

\item{limits}{Should peak boundaries be found on true data (\code{xic}) or
on CentWave scales (\code{scales} = default)?}

\item{...}{Other options passed to \code{\link[MassSpecWavelet]{peakDetectionCWT}}}
}
\value{
A data.table with information on chromatographic peaks
with the following columns:
\itemize{
\item \code{peakID}: Peak name given by \code{\link[MassSpecWavelet]{peakDetectionCWT}}
\item \code{scmin}, \code{scmax}, \code{scpos}: left, right and apex of the peak boundaries from \code{vecint}
\item \code{scale}: scale of the peak given by \code{\link[MassSpecWavelet]{peakDetectionCWT}}
}
}
\description{
Use MassSpecWavelet to detect peaks
}
