% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{get_peak_range}
\alias{get_peak_range}
\title{Use MassSpecWavelet to detect peaks}
\usage{
get_peak_range(xic, direction = c("R", "L"), peak_sel_center)
}
\arguments{
\item{xic}{The XIC as a data.table with the columns:
'scan' and 'i'.}

\item{direction}{Character to choose the right 'R' or left 'L'
side of the peak}

\item{peak_sel_center}{Scan corresponding to the peak apex.}
}
\value{
A numeric value corresponding to the scan on the right \code{R} or
left \code{L} side of the peak.
}
\description{
Use MassSpecWavelet to detect peaks
}
