% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_best_peaks.R
\name{get_peaks_xic}
\alias{get_peaks_xic}
\title{Find peaks usint CentWave}
\usage{
get_peaks_xic(xic_mat, minscan = 3, majpeakonly = FALSE)
}
\arguments{
\item{xic_mat}{a matrix or data.table with the following columns:
- \code{rt} with the retention time
- \code{i} with the intensities}

\item{minscan}{numeric value for the minimum number of scan
inside a peak to keep it}

\item{majpeakonly}{Logical to extract only the majors peaks (\code{TRUE})
or all peaks (\code{FALSE}) from MassSpecWavelet}
}
\description{
Find peaks usint CentWave
}
