% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{load_files}
\alias{load_files}
\title{Create \code{data.table} from files path}
\usage{
load_files(files)
}
\arguments{
\item{files}{A \verb{character vector} of path(s) to \code{.raw} files}
}
\value{
Return a \code{data.table} with the following columns:
\itemize{
\item \code{file_path}: the full file path
\item \code{file_name}: the file base name
\item \code{FileIndex}: a unique \code{integer} to identify the file
\item \code{FileExist}: a \code{logical} to test if the file exist or not
}
}
\description{
Takes \code{.raw} file paths and return a formatted
\code{data.table}.
}
\examples{
load_files(files = rawrr::sampleFilePath())

}
