% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_Spec2Xtract.R
\name{run_Spec2Xtract}
\alias{run_Spec2Xtract}
\title{Wrapper to run targets pipeline}
\usage{
run_Spec2Xtract(
  files_dir,
  cpd_path,
  firstevent,
  prec_ppm,
  minscan,
  rt_limit,
  ppm,
  filter_irel = 0,
  filter_isopurity = 0,
  save_dir,
  ncore = 1
)
}
\arguments{
\item{files_dir}{path to the directory containing the
.raw files}

\item{cpd_path}{path to the table with compound informations}

\item{firstevent}{Should all MSn events be checked individually
(\code{FALSE}) or only the first one (\code{TRUE}).}

\item{prec_ppm}{Mass tolerance to find the precursor (in ppm)}

\item{minscan}{numeric value for the minimum number of scan
inside a peak to keep it}

\item{rt_limit}{limit rt deviation from reference when searching peaks}

\item{ppm}{mass deviation for elemental composition
search in ppm}

\item{filter_irel}{Filter ions based on relative intensity}

\item{filter_isopurity}{Filter ions based on the isolation purity (from 0 to 100)}

\item{save_dir}{Directory to save tables}

\item{ncore}{Number of parallel workers}
}
\description{
This function run a full targets pipeline from
two paths: one to the directory containing .raw
files and one to the compound table.
}
